Remove type information from MappedBuffer. Revert Poke confusion. Allocate descriptor buffers based only on known definitions.
We no longer require a StorableMap neither SomeStorable in the shader
pipeline definition because we compute the buffer sizes from the
KnownDefinitions.

We're just missing validation of material descriptor and shader
descriptor now to get back on track: We describe a shader and the
material we use on render packets with that pipeline must be compatible
with it. See 'Compatible' in Ghengin.Render.Packet.

When we finish 'Compatible', the current MinMax material shouldn't be
compatible with the pipeline which takes two descriptos on set #1 while
MinMax only specifies one property rather than two.