Allow Material in-place edit with HasBindingAt
Previously we weren't able to edit the materials without unsafeCoerce
everywhere. Now, we can use the typeable instance to compare against the
kind of material we want to edit and edit it knowing it is that
material, and furthermore we can recreate the renderpacket because the
compatible constraint isn't lost because we don't create a new material
but rather prove it's a certain material we can edit.

I've implemented the update functions with the exception of updating
Textures, which will come in the next commit