Complete material edit with writing new textures
The design through the use of eqT to be able to edit the material works
out great.

Perhaps I could even do the coercion in the `medit` function. However,
then the user would be rid of the burden of inspecting the type of
material in that render packet, which I'm not sure is a good thing.