Render keys: typeable for pipeline, unique for material
We can now get rid the Hashable instance.

For the render queue, pipelines and materials must be ordered.
To do so, we need a key to uniquely identify pipelines and materials.
Because pipelines are uniquely identified by their type, we can use
their TypeRep from their typeable instance as a key.
Materials are harder to identify uniquely because they depend on the
values used to construct it. To simplify matters, we already required a
call to 'material' to construct a material. That call now simply creates
a 'Unique' and assigns it to each new material.