Fix packaging issue
The issue was the output log file.
Accessing the filesystem (to write the file) made the app crash on start
because the log file was created on start.

After building a thousand different apps across different repositories
and ghengin commits I finally tracked down the issue to the commit which
introduced the logger.