Successfully define a Camera outside of Core in Planets
The culmination of the move to a Core engine that doesn't hardcode
things like Cameras and Lights but rather has them defined through
shaders and render properties.

Eventually, we will want to distribute common logic for things like
Cameras and Lights but these will live *outside* of Core!

There's a lot to be written about the Core of the engine, but in
essence, its benefits resemble those of Haskell's Core intermediate
language.