Linear Types in Render Properties
This is a gigantic win!

We now use backpack to have a renderer-agnostic interface

The definitions in the signature module for the renderer are linear in
the resources that must be tracked linearly

The render properties module now makes use of the renderer agnostic
linear interface only (the propertyAt lens' type had to change a bit to
account for linearity, so we will need a new combinator to use it)

We introduced the reference counting experimental library with great
success, it allows us to have linear data structures containing linear
resources and still be useful (e.g. allow access to one element of the
linear container as a counted reference; which would otherwise be
impossible for containers with linear resources)

We ignore Texture bindings for now, since they need reference counting
of its own and will become simpler when the rest of the code is moved to
use linear types