SUCCESSFULLY COMPILE GHENGIN
That's a complete linear ghengin, with a linear ghengin-core that
defines an abstract graphics API that is then instanced by a linear
ghengin-vulkan.
These are brought together by a quite hacked ghengin.

Unfortunately, in this last step I realized it is probably for the best
to not use linear types in Ghengin, the frontend. Instead, we can wrap
the linear renderer monad which (for now, only somewhat) ensures the
implementation wrt Host and Device resources that must be freed and
manually managed is correct in `UrT`, which would allow Ghengin to be a
normal Monad.

An alternative is to fix the linear-apecs API to actually consider
resource usage, and make the ghengin frontend be usable with linear
types, but such a strict resource managament policy enforcement is
unnecessary in the frontend which can already guarantee the user barely
needs to manage resources manually, and entrust them with those that
they do.

I've been hacking together a TODO list.
Up next I'll be refactoring ghengin-core until it is in a (attempted)
beautiful state, point at which I will begin trying to organize the
project through the issue tracker.
There's also the project of reverting linear types in frontend Ghengin.

When ghengin-core is much better and reworked, it'd be great to make
ghengin more modular, and even remove its dependency on Vulkan.

Then, there's the design document project which attempts to better
define the architecture of the engine -- and then we'll be able to
re-think things like parallelism and concurrency, and further features,
deferred rendering, animations, and much more!