Use cabal.project instead of multilib ghengin.cabal
Move the multiple libraries from the ghengin.cabal file into a separate
projects, each with its .cabal file.

The reason for doing this is that the interaction between HLS and
backpack is broken, so by splitting the packages out we're able to use
HLS at least in the packages that don't have signatures.

We want to eventually merge these projects back to one, but it depends
on backpack x HLS being fixed first.