Fix a texture segfault
Likely caused by Unsafe.toLinear uses allowing a variable to be captured
in a closure and then incorrectly freed before the 'immediateSubmit' is
run. If we move 'destroyBuffer' to after 'immediateSubmit' then it is
fixed, but this should never happen with linear types...

This is in 'textureFromImage'. The 'stagingBuffer' was captured in one
of the commands, then destroyed. It would never have happened if we
didn't use Unsafe.toLinear so liberally! At least it's only in
ghengin-vulkan...